+++ hardhat.config.ts
@@ -1,8 +1,6 @@
 import "module-alias/register";
 
 import "@nomicfoundation/hardhat-chai-matchers";
-import "@nomicfoundation/hardhat-toolbox";
-import "@nomiclabs/hardhat-ethers";
 import "@nomiclabs/hardhat-etherscan";
 import "@openzeppelin/hardhat-upgrades";
 import "@typechain/hardhat";
@@ -88,6 +86,18 @@ const config: HardhatUserConfig = {
       live: true,
       timeout: 1200000, // 20 minutes
     },
+    opbnbtestnet: {
+      url: process.env.ARCHIVE_NODE_opbnbtestnet || "https://opbnb-testnet-rpc.bnbchain.org",
+      chainId: 5611,
+      live: true,
+      accounts: DEPLOYER_PRIVATE_KEY ? [`0x${DEPLOYER_PRIVATE_KEY}`] : [],
+    },
+    opbnbmainnet: {
+      url: process.env.ARCHIVE_NODE_opbnbmainnet || "https://opbnb-mainnet-rpc.bnbchain.org",
+      chainId: 204,
+      live: true,
+      accounts: DEPLOYER_PRIVATE_KEY ? [`0x${DEPLOYER_PRIVATE_KEY}`] : [],
+    },
   },
   etherscan: {
     apiKey: {
@@ -95,6 +105,8 @@ const config: HardhatUserConfig = {
       bsctestnet: process.env.ETHERSCAN_API_KEY || "ETHERSCAN_API_KEY",
       sepolia: process.env.ETHERSCAN_API_KEY || "ETHERSCAN_API_KEY",
       ethereum: process.env.ETHERSCAN_API_KEY || "ETHERSCAN_API_KEY",
+      opbnbtestnet: process.env.ETHERSCAN_API_KEY || "ETHERSCAN_API_KEY",
+      opbnbmainnet: process.env.ETHERSCAN_API_KEY || "ETHERSCAN_API_KEY",
     },
     customChains: [
       {
@@ -127,6 +139,22 @@ const config: HardhatUserConfig = {
         urls: {
           apiURL: "https://api.etherscan.io/api",
           browserURL: "https://etherscan.io",
+        },
+      },
+      {
+        network: "opbnbtestnet",
+        chainId: 5611,
+        urls: {
+          apiURL: `https://open-platform.nodereal.io/${process.env.ETHERSCAN_API_KEY}/op-bnb-testnet/contract/`,
+          browserURL: "https://testnet.opbnbscan.com/",
+        },
+      },
+      {
+        network: "opbnbmainnet",
+        chainId: 204,
+        urls: {
+          apiURL: `https://open-platform.nodereal.io/${process.env.ETHERSCAN_API_KEY}/op-bnb-testnet/contract/`,
+          browserURL: "https://opbnbscan.com/",
         },
       },
     ],
@@ -166,10 +194,10 @@ const config: HardhatUserConfig = {
       },
     ],
     deployments: {
-      bsctestnet: ["node_modules/@venusprotocol/venus-protocol/deployments/bsctestnet"],
-      bscmainnet: ["node_modules/@venusprotocol/venus-protocol/deployments/bscmainnet"],
-      sepolia: [],
-      ethereum: [],
+      bsctestnet: ["node_modules/@venusprotocol/governance-contracts/deployments/bsctestnet"],
+      bscmainnet: ["node_modules/@venusprotocol/governance-contracts/deployments/bscmainnet"],
+      sepolia: ["node_modules/@venusprotocol/governance-contracts/deployments/sepolia"],
+      ethereum: ["node_modules/@venusprotocol/governance-contracts/deployments/ethereum"],
     },
   },
   docgen: {