+++ helpers/deploy/deploymentUtils.ts
@@ -1,7 +1,4 @@
-import bscMainnetDeployments from "@venusprotocol/venus-protocol/deployments/bscmainnet.json";
-import bscTestnetDeployments from "@venusprotocol/venus-protocol/deployments/bsctestnet.json";
 import { ethers, getNamedAccounts } from "hardhat";
-import { HardhatRuntimeEnvironment } from "hardhat/types";
 
 import bscMainnetGovernanceDeployments from "../../deployments/bscmainnet.json";
 import bscTestnetGovernanceDeployments from "../../deployments/bsctestnet.json";
@@ -9,13 +6,18 @@ import { SUPPORTED_NETWORKS } from "./constants";
 
 export const getAcmAdminAccount = async (network: SUPPORTED_NETWORKS): Promise<string> => {
   const { deployer } = await getNamedAccounts();
-  return {
-    bsctestnet: bscTestnetDeployments.contracts.NormalTimelock.address, // NORMAL TIMELOCK
-    bscmainnet: bscMainnetDeployments.contracts.NormalTimelock.address, // NORMAL TIMELOCK
-    sepolia: "0x94fa6078b6b8a26f0b6edffbe6501b22a10470fb", // SEPOLIA MULTISIG
-    ethereum: "0x285960C5B22fD66A736C7136967A3eB15e93CC67", // ETHEREUM MULTISIG
-    hardhat: deployer,
-  }[network];
+  if (network === "hardhat") {
+    return deployer;
+  } else if (network === "sepolia") {
+    return "0x94fa6078b6b8a26f0b6edffbe6501b22a10470fb"; // SEPOLIA MULTISIG
+  } else if (network === "ethereum") {
+    return "0x285960C5B22fD66A736C7136967A3eB15e93CC67"; // ETHEREUM MULTISIG
+  } else if (network === "opbnbtestnet") {
+    return "0xb15f6EfEbC276A3b9805df81b5FB3D50C2A62BDf"; // OPBNBTESTNET MULTISIG
+  }
+
+  const normalTimelock = await ethers.getContract("NormalTimeLock");
+  return normalTimelock.address;
 };
 
 export const getOmnichainProposalSender = async (network: Omit<SUPPORTED_NETWORKS, "ethereum" | "sepolia">) => {
@@ -26,20 +28,6 @@ export const getOmnichainProposalSender = async (network: Omit<SUPPORTED_NETWORK
     return bscTestnetGovernanceDeployments.contracts.OmnichainProposalSender.address;
   }
   return bscMainnetGovernanceDeployments.contracts.OmnichainProposalSender.address;
-};
-
-export const toAddress = async (addressOrAlias: string, hre: HardhatRuntimeEnvironment): Promise<string> => {
-  const { getNamedAccounts } = hre;
-  const { deployments } = hre;
-  if (addressOrAlias.startsWith("0x")) {
-    return addressOrAlias;
-  }
-  if (addressOrAlias.startsWith("account:")) {
-    const namedAccounts = await getNamedAccounts();
-    return namedAccounts[addressOrAlias.slice("account:".length)];
-  }
-  const deployment = await deployments.get(addressOrAlias);
-  return deployment.address;
 };
 
 // eslint-disable-next-line @typescript-eslint/no-empty-function